fleshing out more stuff - compiles

+++ src/flight_controller/inc/scheduler.h
@@ -1,6 +1,6 @@
 #pragma once
 
-void scheduler(uint8_t count);
+void scheduler(int32_t count);
 void taskHandlePcComm(void);
 void taskLed(void);
 void taskBuzzer(void);
